+++ walap_upload/__init__.py
@@ -5,7 +5,7 @@ except ModuleNotFoundError:
 
 PLUGIN_METADATA = {
     'id': 'walap_upload',
-    'version': '0.1.0',
+    'version': '0.2.1',
     'name': 'Walap Upload'
 }